split .h and _main.cc files out
